function ubication corrected
